Update documentation after testing on Linux Mint 21.1; Provide default values.
This change updates the READMEs, especially for server.py, where I noticed
several steps as required which some users may not have thought to check
before-hand. I've also expanded the text to explain a bit more of what's going
on to help the reader be clear.

While getting Authenticated, I got several KeyErrors on environ[] for things
that the documentation says are optional to configure. The options did not seem
to have defaults in server.py, so I've added some.